fix: add correct implementation of runtime messaging bridge
