--- unity-client/Assets/Scripts/MainScripts/DCL/Controllers/Scene/Debug.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 145dfb72e03703a44a5d6d02b1876a36
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
